+++ services/auth/src/main/java/org/sagebionetworks/auth/ResourceAccessManager.java
@@ -14,7 +14,6 @@ import org.sagebionetworks.authutil.AuthenticationException;
 import org.sagebionetworks.authutil.CrowdAuthUtil;
 import org.sagebionetworks.authutil.Session;
 import org.sagebionetworks.authutil.User;
-import org.sagebionetworks.repo.model.AuthorizationConstants;
 import org.sagebionetworks.repo.web.NotFoundException;
